requirements.txt
runpod
requests
Pillow
realesrgan
basicsr
facexlib
numpy
